Corehost with dotnet exe support
We can stash the artifacts from coreclr and corefx and use that to
create a full dotnet toolchain. This is packaged a bit differently, so
you want to either install coreclr+corefx or corehost even though many
of the components are the same.
To actually get the Yocto built components we must make sure that all
the version numbers and RIDs match up. Otherwise corehost will download
missing components from the internet and you will get symbol/so-lib
mismatches.

Reviewed-by: Jens Nordenbro <jens.nordenbro@beijerelectronics.com>